refactor: extract marketplace orchestration and shared DTOs
