+++ api-gateway/src/main/resources/application.yaml
@@ -51,7 +51,7 @@ gateway:
   routes:
     setup:
       id: setup-service
-      uri: http://setup-service:8080
+      uri: lb://setup-service
       paths:
         - /api/setup/**
       strip-prefix: 1
@@ -60,7 +60,7 @@ gateway:
         fallback-uri: forward:/fallback/setup-service
     tenant:
       id: tenant-service
-      uri: http://tenant-service:8080
+      uri: lb://tenant-service
       paths:
         - /api/tenant/**
       strip-prefix: 1
@@ -69,7 +69,7 @@ gateway:
         fallback-uri: forward:/fallback/tenant-service
     catalog:
       id: catalog-service
-      uri: http://catalog-service:8080
+      uri: lb://catalog-service
       paths:
         - /api/catalog/**
       strip-prefix: 1
@@ -78,7 +78,7 @@ gateway:
         fallback-uri: forward:/fallback/catalog-service
     subscription:
       id: subscription-service
-      uri: http://subscription-service:8080
+      uri: lb://subscription-service
       paths:
         - /api/subscription/**
       strip-prefix: 1
@@ -87,7 +87,7 @@ gateway:
         fallback-uri: forward:/fallback/subscription-service
     billing:
       id: billing-service
-      uri: http://billing-service:8080
+      uri: lb://billing-service
       paths:
         - /api/billing/**
       strip-prefix: 1
