+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/subscription/dto/EnvironmentIdentifierDto.java
@@ -1,0 +1,7 @@
+package com.ejada.common.marketplace.subscription.dto;
+
+/** Identifier values we may return to marketplace after provisioning. */
+public record EnvironmentIdentifierDto(
+        String identifierCd,
+        String identifierValue) {
+}
