+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/subscription/dto/ProductPropertyDto.java
@@ -1,0 +1,8 @@
+package com.ejada.common.marketplace.subscription.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ProductPropertyDto(
+        @NotBlank String propertyCd,
+        @NotBlank String propertyValue) {
+}
