+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/subscription/dto/SubscriptionUpdateType.java
@@ -1,0 +1,8 @@
+package com.ejada.common.marketplace.subscription.dto;
+
+public enum SubscriptionUpdateType {
+    SUSPENDED,
+    RESUMED,
+    TERMINATED,
+    EXPIRED
+}
