+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/token/TokenHashing.java
@@ -1,0 +1,29 @@
+package com.ejada.common.marketplace.token;
+
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+/** Utility helpers for hashing sensitive marketplace tokens before persistence. */
+public final class TokenHashing {
+
+    private TokenHashing() {
+    }
+
+    public static String sha256(final String value) {
+        if (value == null) {
+            return null;
+        }
+        try {
+            MessageDigest digest = MessageDigest.getInstance("SHA-256");
+            byte[] hash = digest.digest(value.getBytes(StandardCharsets.UTF_8));
+            StringBuilder builder = new StringBuilder(hash.length * 2);
+            for (byte b : hash) {
+                builder.append(String.format("%02x", b));
+            }
+            return builder.toString();
+        } catch (NoSuchAlgorithmException e) {
+            return null;
+        }
+    }
+}
