--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/AdminUserInfoDto.java
@@ -1,11 +1,0 @@
-package com.ejada.subscription.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record AdminUserInfoDto(
-    @NotBlank String adminUserName,
-    String preferredLang,    // EN | AR
-    @NotBlank String mobileNo,
-    @Email @NotBlank String email
-) { }
