--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/CustomerInfoDto.java
@@ -1,16 +1,0 @@
-package com.ejada.subscription.dto;
-
-import jakarta.validation.constraints.Email;
-
-public record CustomerInfoDto(
-    String customerNameEn,
-    String customerNameAr,
-    String customerType,      // e.g., COMPANY
-    String crNumber,
-    String countryCd,
-    String cityCd,
-    String addressEn,
-    String addressAr,
-    @Email String email,
-    String mobileNo
-) { }
