--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/ReceiveSubscriptionUpdateRq.java
@@ -1,10 +1,0 @@
-package com.ejada.subscription.dto;
-
-import jakarta.validation.constraints.NotNull;
-
-/** Update event from marketplace: suspend/resume/terminate/expire. */
-public record ReceiveSubscriptionUpdateRq(
-    @NotNull Long subscriptionId,
-    @NotNull Long customerId,
-    @NotNull SubscriptionUpdateType  subscriptionUpdateType // SUSPENDED | RESUMED | TERMINATED | EXPIRED
-) { }
