--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/SubscriptionUpdateType.java
@@ -1,3 +1,0 @@
-package com.ejada.subscription.dto;
-
-public enum SubscriptionUpdateType { SUSPENDED, RESUMED, TERMINATED, EXPIRED }
