+++ tenant-platform/subscription-service/src/main/java/com/ejada/subscription/mapper/SubscriptionEnvironmentIdentifierMapper.java
@@ -1,6 +1,6 @@
 package com.ejada.subscription.mapper;
 
-import com.ejada.subscription.dto.EnvironmentIdentifierDto;
+import com.ejada.common.marketplace.subscription.dto.EnvironmentIdentifierDto;
 import com.ejada.subscription.model.Subscription;
 import com.ejada.subscription.model.SubscriptionEnvironmentIdentifier;
 import org.mapstruct.Context;
